requirements.txt
python-docx
gradio
pandas
langchain     
chromadb      
gradio
python-docx